Refactor listings page testing into a new folder
