+++ components/tabs/Listings/NewListingTab.jsx
@@ -1,4 +1,3 @@
-import { supabase } from "@/utils/supabase";
 import {
   ScrollView,
   StyleSheet,
